Refactored a variable from DevicePortsFactory
It was required to change the type to size_t as it is a size variable and no decrement operations are performed on it

+++ connectioninputparser.cpp
@@ -92,7 +92,7 @@ bool ConnectionInputParser::_parseInput()
 
     if (!c_ErrorsOccurred)
     {
-        assert(mpDevicePortsFactory->getCreatedDevicePortsCount() == static_cast<int>(mCablePartNumbersEntries.size()) * c_DevicesPerConnectionInputRowCount); // 1 cable, two connected devices
+        assert(mpDevicePortsFactory->getCreatedDevicePortsCount() == mCablePartNumbersEntries.size() * c_DevicesPerConnectionInputRowCount); // 1 cable, two connected devices
     }
 
     return c_ErrorsOccurred;
